service/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
gunicorn==20.0.4

